+++ tests/Integration/Filters/Processors/SearchProcessorTest.php
@@ -6,6 +6,7 @@ use Aedart\Filters\Processors\SearchProcessor;
 use Aedart\Testing\Helpers\ConsoleDebugger;
 use Aedart\Tests\Helpers\Dummies\Database\Models\Category;
 use Aedart\Tests\TestCases\Filters\FiltersTestCase;
+use Illuminate\Validation\ValidationException;
 
 /**
  * SearchProcessorTest
@@ -21,7 +22,7 @@ class SearchProcessorTest extends FiltersTestCase
     /**
      * @test
      *
-     * @throws \Illuminate\Validation\ValidationException
+     * @throws ValidationException
      */
     public function canBuildSearchQuery()
     {
@@ -60,7 +61,7 @@ class SearchProcessorTest extends FiltersTestCase
     /**
      * @test
      *
-     * @throws \Illuminate\Validation\ValidationException
+     * @throws ValidationException
      */
     public function canSearchForZero()
     {
@@ -94,5 +95,49 @@ class SearchProcessorTest extends FiltersTestCase
         ConsoleDebugger::output($sql);
 
         $this->assertNotEmpty($sql, 'Query was not built');
+    }
+
+    /**
+     * @test
+     *
+     * @return void
+     *
+     * @throws ValidationException
+     */
+    public function canSearchUsingCustomCallback(): void
+    {
+        $key = 'search';
+        $processors = [
+            $key => SearchProcessor::make()
+                ->columns(function ($query, $search) {
+                    return $query->where('my_column', '=', $search);
+                })
+        ];
+
+        $builder = $this->makeGenericBuilder($processors, $this->makeRequest(
+            'https://some-url.org/api/v1',
+            'GET',
+            [
+                'search' => 'Lina Hollands'
+            ]
+        ));
+
+        $built = $builder->build();
+        $this->assertTrue($built->has($key), 'Filter was not built');
+
+        // --------------------------------------------------------------- //
+
+        $filters = $built->get($key);
+        $this->assertCount(1, $filters);
+
+        // Apply filter and assert that sql can be generated...
+        $filter = $filters[0];
+        $query = $filter->apply(Category::query());
+
+        $sql = $query->toSql();
+        ConsoleDebugger::output($sql);
+
+        $this->assertNotEmpty($sql, 'Query was not built');
+        $this->assertStringContainsString('where (`my_column` = ?)', $sql);
     }
 }