+++ CHANGELOG.md
@@ -13,7 +13,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * API Request abstractions, in the Http Api package.
 * `HttpCaching` concern in `ApiResource`, which can make it easier to set Http Cache Control headers.
 * `HasArbitraryData` interface and a default implementation in `ArbitraryData` concern, as part of the utils package. 
-* `sync()` method added for `FileStream`. [#105](https://github.com/aedart/athenaeum/issues/105)
+* `sync()` method added for `FileStream`. [#105](https://github.com/aedart/athenaeum/issues/105).
+* `BaseSearchQuery` and `BaseSortingQuery` abstractions for custom filtering queries via `SearchFilter` or `SortFilter`, in the filters package.
+* `Database` utility component, in the database package.
+* Query `Joins` concern, in the database package.
+* `Prefixing` concern, in the database package.
 * Test `Response` utility.
 
 ### Changed
@@ -22,10 +26,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Minimum required PHP version changed to `v8.1`.
 * Adapted CI environment to test PHP `v8.1` and `v8.2`.
-* `ApiResourceServiceProvider` changed to be an aggregate service provider that automatically registers `ETagsServiceProvider`. 
+* `ApiResourceServiceProvider` changed to be an aggregate service provider that automatically registers `ETagsServiceProvider`.
+* `SearchFilter` no longer applies unnecessary query constraint (_the first comparison constraint_).
 
 **Non-breaking Changes**
 
+* `SearchFilter` and `SearchProcessor` now support custom search callbacks. [#129](https://github.com/aedart/athenaeum/issues/129).
+* `SortFilter` and `SortingProcessor` now support custom sorting callbacks.
 * `getResourceKeyName()` in `ApiResource` now throws `LogicException`, if unable to determine resource's identifier key name.
 * `hash()` method can now accept options for the specified hashing algorithm. [#106](https://github.com/aedart/athenaeum/issues/106)
 * Temporary and public URL tests for database adapter are forced to evaluate to true. Original tests marked them as skipped, because features are not supported.